path

+++ Get-LaptopSoftpaq.ps1
@@ -15,27 +15,27 @@ function Get-LaptopSoftpaq {
     $progressPreference = "SilentlyContinue"
 
     $Bios = (Get-ComputerInfo).BiosManufacturer
-    if ($Bios -eq "HP") {
+    $CmslPath = Test-Path -Path "C:\Program Files\WindowsPowerShell\HP.CMSL.UninstallerData"
+    if ($Bios -eq "HP" -and $CmslPath -notmatch "True") {
         Invoke-WebRequest -Uri "https://hpia.hpcloud.hp.com/downloads/cmsl/hp-cmsl-1.6.7.exe" -OutFile "C:\Windows\Temp\HpModule.exe"
         Start-Process -FilePath "C:\Windows\Temp\HpModule.exe" -Wait -ArgumentList "/verysilent /norestart"
         Remove-Item -Path "C:\Windows\Temp\HpModule.exe" -Force
         #Start-Process -FilePath "https://developers.hp.com/hp-client-management/doc/client-management-script-library"
     
-        Start-Job -Name "Get-LaptopSoftpaq" -ScriptBlock {
-            $TestPath = Test-Path -Path "C:\SOFTPAQ"
-            if ($TestPath -match "False") {
-                New-Item -Name "SOFTPAQ" -ItemType Directory -Path "C:\" -Force
-            }
-            $OsVer = (Get-ComputerInfo).OSDisplayVersion
-            $CsModel = (Get-ComputerInfo).CsModel
-            New-Item -Name $CsModel -ItemType Directory -Path "C:\SOFTPAQ\" -Force
-            $Path = "C:\SOFTPAQ\$CsModel\"
-            Invoke-Item -Path $Path
-            New-HPDriverPack -OSVer $OsVer -Path $Path -RemoveOlder -Overwrite    
+        $TestPath = Test-Path -Path "C:\SOFTPAQ"
+        if ($TestPath -match "False") {
+            New-Item -Name "SOFTPAQ" -ItemType Directory -Path "C:\" -Force
         }
-        Get-Job | Wait-Job -Timeout 300
+        $OsVer = (Get-ComputerInfo).OSDisplayVersion
+        $CsModel = (Get-ComputerInfo).CsModel
+        New-Item -Name $CsModel -ItemType Directory -Path "C:\SOFTPAQ\" -Force
+        $Path = "C:\SOFTPAQ\$CsModel\"
+        Invoke-Item -Path $Path
+        New-HPDriverPack -OSVer $OsVer -Path $Path -RemoveOlder -Overwrite    
     }
+
 }
+
 Get-LaptopSoftpaq
 
 Exit
